rename auth module class.
