+++ packages/auth/src/index.ts
@@ -1,5 +1,5 @@
-import { MoralisAuth } from './MoralisAuth';
+import { Auth } from './Auth';
 
-export * from './MoralisAuth';
+export * from './Auth';
 
-export default MoralisAuth;
+export default Auth;
