+++ packages/auth/src/methods/requestMessage.ts
@@ -2,7 +2,7 @@ import { SolAddressish, SolNetworkish, SolAddress, SolNetwork } from '@moraliswe
 import { EndpointResolver } from '@moralisweb3/api-utils';
 import Core, { AuthErrorCode, MoralisAuthError } from '@moralisweb3/common-core';
 import { EvmAddress, EvmAddressish, EvmChain, EvmChainish } from '@moralisweb3/common-evm-utils';
-import { BASE_URL } from '../MoralisAuth';
+import { BASE_URL } from '../Auth';
 import { initializeChallengeEvm, initializeChallengeSol } from '../resolvers';
 import { AuthNetworkType } from '../utils/AuthNetworkType';
